:recycle: Rename files
