+++ lib/domain/menstruation/menstruation_store.dart
@@ -6,7 +6,7 @@ import 'package:pilll/components/organisms/calendar/utility.dart';
 import 'package:pilll/domain/menstruation/menstruation_card_state.dart';
 import 'package:pilll/domain/menstruation/menstruation_history_card_state.dart';
 import 'package:pilll/entity/menstruation.dart';
-import 'package:pilll/native/global_method_channel.dart';
+import 'package:pilll/native/menstruation.dart';
 import 'package:pilll/service/diary.dart';
 import 'package:pilll/service/menstruation.dart';
 import 'package:pilll/service/pill_sheet.dart';
