+++ lib/domain/settings/setting_page_store.dart
@@ -3,7 +3,7 @@ import 'dart:io';
 
 import 'package:pilll/database/batch.dart';
 import 'package:pilll/entity/setting.dart';
-import 'package:pilll/native/global_method_channel.dart';
+import 'package:pilll/native/menstruation.dart';
 import 'package:pilll/service/pill_sheet.dart';
 import 'package:pilll/service/pill_sheet_group.dart';
 import 'package:pilll/service/pill_sheet_modified_history.dart';
